+++ app/core/__init__.py
@@ -2,6 +2,6 @@ from app.core.config import settings
 from app.core.database import engine, Base, get_db
 from app.core.celery import celery_app
 from app.core.logger import logger
-from app.core.redis import get_redis
+from app.core.redis import RedisClient
 
-__all__ = ['settings', 'engine', 'Base', 'logger', 'get_db', 'celery_app', 'get_redis']
+__all__ = ['settings', 'engine', 'Base', 'logger', 'get_db', 'celery_app', 'RedisClient']