+++ app/core/redis.py
@@ -1,13 +1,24 @@
 import redis.asyncio as redis
-from app.core import settings
+from typing import Optional
+from app.core.config import settings
 
-_redis_client = redis.from_url(
-    settings.REDIS_URL,
-    decode_responses = True,
-    socket_connect_timeout = 5,
-    socket_timeout = 5
-)
+class RedisClient:
+    _client: Optional[redis.Redis] = None
 
+    @classmethod
+    async def get_client(cls) -> Optional[redis.Redis]:
+        if cls._client is None:
+            cls._client = redis.from_url(
+                url = settings.REDIS_URL,
+                decode_responses = True,
+                max_connections = 10
+            )
+            await cls._client.ping()
+        return cls._client
 
-async def get_redis() -> redis.Redis: # dependency for getting a redis client
-    return _redis_client
+    @classmethod
+    async def close(cls) -> None:
+        if cls._client:
+            await cls._client.close()
+            cls._client = None
+