+++ app/dependencies.py
@@ -1,6 +1,8 @@
+from uuid import UUID
 from fastapi import Depends
 from fastapi.security import OAuth2PasswordBearer
 from sqlalchemy.ext.asyncio import AsyncSession
+from app.models import User
 from app.core import get_db
 from app.exceptions import user as user_exc
 from app.services import AuthService, UserService, VerificationService, GitHubService, GitHubFeedService
@@ -22,7 +24,7 @@ async def get_github_feed_service(service: GitHubService = Depends(get_github_se
     return GitHubFeedService(service)
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl = '/v1/login')
-async def get_current_user(token: str = Depends(oauth2_scheme), service: UserService = Depends(get_user_service)):
+async def get_current_user(token: str = Depends(oauth2_scheme), service: UserService = Depends(get_user_service)) -> User:
     payload = decode_jwt_token(token)
     if not payload:
         raise user_exc.CouldNotValidateCredentialsError()
@@ -31,3 +33,11 @@ async def get_current_user(token: str = Depends(oauth2_scheme), service: UserSer
         raise user_exc.CouldNotValidateCredentialsError()
     return await service.get_by_id(user_id)
 
+async def get_current_user_id(token: str = Depends(oauth2_scheme)) -> UUID: 
+    payload = decode_jwt_token(token)
+    if not payload:
+        raise user_exc.CouldNotValidateCredentialsError()
+    user_id = payload.get('sub')
+    if not user_id:
+        raise user_exc.CouldNotValidateCredentialsError()
+    return user_id