+++ app/main.py
@@ -7,15 +7,16 @@ from fastapi.responses import JSONResponse
 from app.api import router
 from app.middlewares import LoggingMiddleware
 from app.middlewares import RateLimit
-from app.core import engine, Base, logger, get_redis
+from app.core import engine, Base, logger, RedisClient
 
 @asynccontextmanager
 async def lifespan(app: FastAPI) -> AsyncGenerator:
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.create_all)
-        await (await get_redis()).ping()
+    await (await RedisClient.get_client()).ping()
     logger.info('Prism started')
     yield
+    await (await RedisClient.get_client()).close()
 
 app = FastAPI(lifespan = lifespan)